swissairdry_komplettpaket/docker-base-api/app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pandas==2.0.1
jinja2==3.1.2
aiofiles==23.1.0
requests==2.29.0
python-dotenv==1.0.0